+++ frontend/src/layouts/AccountLayout.vue
@@ -1,0 +1,12 @@
+<script setup>
+// import { EventSubNav } from '@/components';
+</script>
+
+<template>
+  <div class="layout-with-subnav main-content-container">
+    SubNav
+    <div class="sub-content-container">
+      <router-view />
+    </div>
+  </div>
+</template>
