+++ frontend/src/layouts/index.js
@@ -1,2 +1,3 @@
+export { default as Account } from "./AccountLayout.vue";
 export { default as Event } from "./EventLayout.vue";
 export { default as Home } from "./HomeLayout.vue";
