+++ frontend/src/router/event.routes.js
@@ -9,6 +9,6 @@ export default {
     { path: "new", name: "new", component: () => import("@/views/events/New") },
     { path: "past", name: "past", component: () => import("@/views/events/Past") },
     { path: ":uid", name: "staged", component: () => import("@/views/events/Staged.vue"), props: true },
-    { path: ":uid/edit", name: "editEvent/:uid", component: () => import("@/views/events/Edit.vue"), props: true },
+    { path: ":uid/edit", name: "editEvent/:uid", component: () => import("@/views/events/EditEvent.vue"), props: true },
   ]
 }
