+++ frontend/src/views/index.js
@@ -1,4 +1,4 @@
 export { default as Welcome } from "./HomeView.vue";
-export { default as User } from "./UserView.vue";
+export { default as AccountView } from "./account/AccountView.vue";
 
 export { default as Manager } from "./events/Manager.vue";
